modified so user could put in less than 1 hour

+++ lib/dialog/edit_goal_dialog.dart
@@ -27,8 +27,8 @@ class _EditGoalDialogState extends State<EditGoalDialog>{
     return result;
   }
 
-  int getTimeValue() {
-    final result = int.tryParse(timeController.value.text);
+  double getTimeValue() {
+    final result = double.tryParse(timeController.value.text);
     if (result == null) return 0;
     return result;
   }
