+++ lib/model/edit_goal_directive.dart
@@ -3,7 +3,7 @@ class EditGoal {
   String title;
   String? description;
   double costInDollars = 0.0;
-  int timeInHours = 0;
+  double timeInHours = 0;
   bool complete = false;
 
   EditGoal(this.title, this.description, this.costInDollars,
